dashboard/backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
pymongo==4.3.3
psycopg2-binary==2.9.5
python-dotenv==1.0.0
requests==2.28.2